requirements.txt
# requirements.txt

arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-tailwind==3.8.0
djangorestframework==3.14.0
dnspython==2.5.0
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
psycopg2-binary==2.9.9
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
urllib3==2.2.0
gunicorn==21.2.0
whiteNoise>=6.6.0
redis>=5.0.1
gevent>=24.2.1
libgravatar>=1.0.4
